Refactor ICOGenData
Encountered a number of bugs related to stale state, so refactored to reduce the amount of state tracked. Especially in regards to tracking derived images and when they needed to be updated. The active images are also no longer tracked. As derived images are now no longer automatically generated and only created when requested, there's no need to manually track which sizes are actively used. Should now be much easier to maintain.