Fix Windows Explorer bug by reversing image order
According to the ICO file-format standard, image-order doesn't matter except for images in BMP format and their masks. They can be encoded into the file in any order, although convention states they should be in the same order as defined in the image directory. As it turns out, at least Windows Explorer does care about image order. If the images are encoded from the smallest to largest image, the icon will be blurry in Explorer, as it'll only use and scale the smallest image. If the order is reversed from largest to smallest image, Explorer has no trouble reading all the image sizes properly.